move to schema.py and test_schema.py
